Modified the description of permit status codes and how they are displayed.
