+++ frontend/src/actionCreators/staticContentActionCreator.js
@@ -149,6 +149,26 @@ export const fetchExpectedDocumentStatusOptions = () => (dispatch) => {
     });
 };
 
+export const fetchPermitStatusOptions = () => (dispatch) => {
+  dispatch(request(reducerTypes.GET_PERMIT_STATUS_OPTIONS));
+  dispatch(showLoading("modal"));
+  return axios
+    .get(`${ENVIRONMENT.apiUrl}${API.PERMIT()}/status-codes`, createRequestHeader())
+    .then((response) => {
+      dispatch(success(reducerTypes.GET_PERMIT_STATUS_OPTIONS));
+      dispatch(staticContentActions.storePermitStatusOptions(response.data));
+      dispatch(hideLoading("modal"));
+    })
+    .catch((err) => {
+      notification.error({
+        message: err.response ? err.response.data.error.message : String.ERROR,
+        duration: 10,
+      });
+      dispatch(error(reducerTypes.GET_PERMIT_STATUS_OPTIONS));
+      dispatch(hideLoading("modal"));
+    });
+};
+
 export const setOptionsLoaded = () => (dispatch) => {
   dispatch(staticContentActions.loadedOptions(true));
 };
