+++ frontend/src/actions/permitActions.js
@@ -4,8 +4,3 @@ export const storePermits = (payload) => ({
   type: actionTypes.STORE_PERMITS,
   payload,
 });
-
-export const storePermitStatusOptions = (payload) => ({
-  type: actionTypes.STORE_PERMIT_STATUS_OPTIONS,
-  payload,
-});
