+++ frontend/src/components/dashboard/minesHomePage/Dashboard.js
@@ -17,6 +17,7 @@ import {
   fetchMineTenureTypes,
   fetchMineDisturbanceOptions,
   fetchMineCommodityOptions,
+  fetchPermitStatusOptions,
   setOptionsLoaded,
 } from "@/actionCreators/staticContentActionCreator";
 import { fetchPartyRelationshipTypes } from "@/actionCreators/partiesActionCreator";
@@ -110,6 +111,7 @@ export class Dashboard extends Component {
       this.props.fetchMineDisturbanceOptions();
       this.props.fetchMineCommodityOptions();
       this.props.fetchPartyRelationshipTypes();
+      this.props.fetchPermitStatusOptions();
       this.props.setOptionsLoaded();
     }
   }
@@ -411,6 +413,7 @@ const mapDispatchToProps = (dispatch) =>
       fetchMineTenureTypes,
       fetchMineCommodityOptions,
       fetchMineDisturbanceOptions,
+      fetchPermitStatusOptions,
       openModal,
       closeModal,
       setOptionsLoaded,
