+++ frontend/src/components/mine/Permit/MinePermitInfo.js
@@ -5,11 +5,7 @@ import PropTypes from "prop-types";
 import CustomPropTypes from "@/customPropTypes";
 import * as Permission from "@/constants/permissions";
 import AuthorizationWrapper from "@/components/common/wrappers/AuthorizationWrapper";
-import {
-  fetchPermits,
-  fetchPermitStatusOptions,
-  createPermit,
-} from "@/actionCreators/permitActionCreator";
+import { fetchPermits, createPermit } from "@/actionCreators/permitActionCreator";
 import { Icon, Button } from "antd";
 import MinePermitTable from "@/components/mine/Permit/MinePermitTable";
 import * as ModalContent from "@/constants/modalContent";
@@ -27,7 +23,6 @@ const propTypes = {
   closeModal: PropTypes.func.isRequired,
   createPermit: PropTypes.func.isRequired,
   fetchPermits: PropTypes.func.isRequired,
-  fetchPermitStatusOptions: PropTypes.func.isRequired,
 };
 
 const defaultProps = {
@@ -38,7 +33,6 @@ const defaultProps = {
 export class MinePermitInfo extends Component {
   componentWillMount() {
     this.props.fetchPermits({ mine_guid: this.props.mine.guid });
-    this.props.fetchPermitStatusOptions();
   }
 
   openAddPermitModal = (event, onSubmit, title) => {
@@ -118,7 +112,6 @@ const mapDispatchToProps = (dispatch) =>
   bindActionCreators(
     {
       fetchPermits,
-      fetchPermitStatusOptions,
       createPermit,
     },
     dispatch
