+++ frontend/src/reducers/permitReducer.js
@@ -1,6 +1,5 @@
 import * as actionTypes from "@/constants/actionTypes";
 import { PERMITS } from "@/constants/reducerTypes";
-import { createItemMap, createItemIdsArray, createDropDownList } from "@/utils/helpers";
 
 const initialState = {
   permits: [],
@@ -13,18 +12,11 @@ const permitReducer = (state = initialState, action) => {
         ...state,
         permits: action.payload,
       };
-    case actionTypes.STORE_PERMIT_STATUS_OPTIONS:
-      return {
-        ...state,
-        statusCodes: action.payload,
-      };
     default:
       return state;
   }
 };
 
 export const getPermits = (state) => state[PERMITS].permits;
-export const getPermitStatusOptions = (state) =>
-  createDropDownList(state[PERMITS].statusCodes, "description", "permit_status_code");
 
 export default permitReducer;
